+++ apps/forms.py
@@ -2,7 +2,8 @@ from typing import Any
 
 from django.forms import ModelForm
 
-from apps.models import Application, ResourceStub, ResponseStub
+from apps.models import Application, RequestStub, ResourceStub, ResponseStub
+from apps.wigdets import Editor
 
 
 class ResourceStubForm(ModelForm):
@@ -43,3 +44,31 @@ class ResponseStubForm(ModelForm):
 
     class Meta:
         model: ResponseStub
+        fields = '__all__'
+        widgets = {
+            'headers': Editor(attrs={'style': 'width: 90%; height: 100%;'}),
+            'body': Editor(attrs={'style': 'width: 90%; height: 100%;'}),
+        }
+
+
+class WebHookRequestForm(ModelForm):
+    def __init__(self, *args: Any, **kwargs: Any) -> None:
+        super().__init__(*args, **kwargs)
+        self.fields['body'].strip = False
+
+        if args:
+            return
+        if data := kwargs.get('initial'):
+            self.fields['application'].queryset = Application.objects.filter(pk=data.get('application'))
+        elif self.instance and hasattr(self.instance, 'application'):
+            self.fields['application'].queryset = Application.objects.filter(pk=self.instance.application.pk)
+        else:
+            self.fields['application'].queryset = Application.objects.none()
+
+    class Meta:
+        model: RequestStub
+        fields = '__all__'
+        widgets = {
+            'headers': Editor(attrs={'style': 'width: 90%; height: 100%;'}),
+            'body': Editor(attrs={'style': 'width: 90%; height: 100%;'}),
+        }