+++ changelog.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.0] - 2022-07-11
 ### Added
 - Global proxy functionality.
 - Ability to stub the specific resource from proxy logging.
@@ -117,7 +118,8 @@ environment variables.
 - This changelog file.
 - Initial Readme file. 
 
-[Unreleased]: https://github.com/s-nagaev/stubborn/compare/v0.4.1...HEAD
+[Unreleased]: https://github.com/s-nagaev/stubborn/compare/v0.5.0...HEAD
+[0.5.0]: https://github.com/s-nagaev/stubborn/compare/v0.4.1...v0.5.0
 [0.4.1]: https://github.com/s-nagaev/stubborn/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/s-nagaev/stubborn/compare/v0.3.3...v0.4.0
 [0.3.3]: https://github.com/s-nagaev/stubborn/compare/v0.3.2...v0.3.3